ADODB->Devblocks
* [Platform/Performance] Removed the dependency on ADODB and migrated all database interaction to Devblocks2.  This reduces the number of open files, reduces memory, and improves compile times.  We didn't need a multi-platofrm abstraction layer since our MySQL support is exclusive, and the wrappers around the native MySQL drivers had a performance hit.

* [Platform] All database scripts have been migrated to MySQL-compatible SQL and all references to ADODB's DataDict have been phased out.

* [Platform/Performance] Peak memory usage has been reduced by proper usage of mysql_free_result().

* [Platform] Refactored out the deprecated 'db_type' param of DevblocksSearchCriteria() since it was only used in 2 out of a thousand use cases.  Blobs no longer need special handling with ADODB removed.

+++ features/cerberusweb.watchers/patches/1.0.0.php
@@ -1,90 +1,35 @@
 <?php
-/***********************************************************************
-| Cerberus Helpdesk(tm) developed by WebGroup Media, LLC.
-|-----------------------------------------------------------------------
-| All source code & content (c) Copyright 2007, WebGroup Media LLC
-|   unless specifically noted otherwise.
-|
-| This source code is released under the Cerberus Public License.
-| The latest version of this license can be found here:
-| http://www.cerberusweb.com/license.php
-|
-| By using this software, you acknowledge having read this license
-| and agree to be bound thereby.
-| ______________________________________________________________________
-|	http://www.cerberusweb.com	  http://www.webgroupmedia.com/
-***********************************************************************/
-/*
- * IMPORTANT LICENSING NOTE from your friends on the Cerberus Helpdesk Team
- * 
- * Sure, it would be so easy to just cheat and edit this file to use the 
- * software without paying for it.  But we trust you anyway.  In fact, we're 
- * writing this software for you! 
- * 
- * Quality software backed by a dedicated team takes money to develop.  We 
- * don't want to be out of the office bagging groceries when you call up 
- * needing a helping hand.  We'd rather spend our free time coding your 
- * feature requests than mowing the neighbors' lawns for rent money. 
- * 
- * We've never believed in encoding our source code out of paranoia over not 
- * getting paid.  We want you to have the full source code and be able to 
- * make the tweaks your organization requires to get more done -- despite 
- * having less of everything than you might need (time, people, money, 
- * energy).  We shouldn't be your bottleneck.
- * 
- * We've been building our expertise with this project since January 2002.  We 
- * promise spending a couple bucks [Euro, Yuan, Rupees, Galactic Credits] to 
- * let us take over your shared e-mail headache is a worthwhile investment.  
- * It will give you a sense of control over your in-box that you probably 
- * haven't had since spammers found you in a game of "E-mail Address 
- * Battleship".  Miss. Miss. You sunk my in-box!
- * 
- * A legitimate license entitles you to support, access to the developer 
- * mailing list, the ability to participate in betas and the warm fuzzy 
- * feeling of feeding a couple obsessed developers who want to help you get 
- * more done than 'the other guy'.
- *
- * - Jeff Standen, Mike Fogg, Brenan Cavish, Darren Sugita, Dan Hildebrandt
- * 		and Joe Geck.
- *   WEBGROUP MEDIA LLC. - Developers of Cerberus Helpdesk
- */
-
 $db = DevblocksPlatform::getDatabaseService();
-$datadict = NewDataDictionary($db); /* @var $datadict ADODB_DataDict */ // ,'mysql' 
-
-$tables = $datadict->MetaTables();
-$tables = array_flip($tables);
+$tables = $db->metaTables();
 
 // `worker_mail_forward` ========================
 if(!isset($tables['worker_mail_forward'])) {
-	$flds ="
-		id I4 DEFAULT 0 NOTNULL PRIMARY,
-		worker_id I4 DEFAULT 0 NOTNULL,
-		group_id I4 DEFAULT 0 NOTNULL,
-		bucket_id I4 DEFAULT -1 NOTNULL,
-		email C(128) DEFAULT '' NOTNULL,
-		event C(3) DEFAULT '' NOTNULL
+	$sql = "
+		CREATE TABLE IF NOT EXISTS worker_mail_forward (
+			id INT UNSIGNED DEFAULT 0 NOT NULL,
+			worker_id INT UNSIGNED DEFAULT 0 NOT NULL,
+			group_id INT UNSIGNED DEFAULT 0 NOT NULL,
+			bucket_id INT UNSIGNED DEFAULT -1 NOT NULL,
+			email VARCHAR(128) DEFAULT '' NOT NULL,
+			event VARCHAR(3) DEFAULT '' NOT NULL,
+			PRIMARY KEY (id)
+		) ENGINE=MyISAM;
 	";
-	$sql = $datadict->CreateTableSQL('worker_mail_forward', $flds);
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute($sql);	
 }
 
-$columns = $datadict->MetaColumns('worker_mail_forward');
-$indexes = $datadict->MetaIndexes('worker_mail_forward',false);
+list($columns, $indexes) = $db->metaTable('worker_mail_forward');
 
 if(!isset($indexes['worker_id'])) {
-	$sql = $datadict->CreateIndexSQL('worker_id','worker_mail_forward','worker_id');
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute('ALTER TABLE worker_mail_forward ADD INDEX worker_id (worker_id)');
 }
 
 if(!isset($indexes['group_id'])) {
-	$sql = $datadict->CreateIndexSQL('group_id','worker_mail_forward','group_id');
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute('ALTER TABLE worker_mail_forward ADD INDEX group_id (group_id)');
 }
 
 if(!isset($indexes['bucket_id'])) {
-	$sql = $datadict->CreateIndexSQL('bucket_id','worker_mail_forward','bucket_id');
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute('ALTER TABLE worker_mail_forward ADD INDEX bucket_id (bucket_id)');
 }
 
 /*
@@ -95,14 +40,16 @@ $sql = "SELECT DISTINCT wmf.worker_id
 	FROM worker_mail_forward wmf 
 	LEFT JOIN worker w ON (w.id=wmf.worker_id) 
 	WHERE w.id IS NULL";
-$rs = $db->Execute($sql);
 
-while(!$rs->EOF) {
-	$sql = sprintf("DELETE FROM worker_mail_forward WHERE worker_id = %d",
-		$rs->fields['worker_id']
-	);
-	$db->Execute($sql);
-	$rs->MoveNext();
+if(false !== ($rs = $db->Execute($sql))) {
+	while($row = mysql_fetch_assoc($rs)) {
+		$sql = sprintf("DELETE FROM worker_mail_forward WHERE worker_id = %d",
+			$row['worker_id']
+		);
+		$db->Execute($sql);
+	}
+	
+	mysql_free_result($rs);
 }
 
 // ===========================================================================
@@ -118,17 +65,19 @@ $db->Execute($sql);
 // Add a table for new worker notification filters
 
 if(!isset($tables['watcher_mail_filter'])) {
-	$flds ="
-		id I4 DEFAULT 0 NOTNULL PRIMARY,
-		pos I2 DEFAULT 0 NOTNULL,
-		name C(128) DEFAULT '' NOTNULL,
-		created I4 DEFAULT 0 NOTNULL,
-		worker_id I4 DEFAULT 0 NOTNULL,
-		criteria_ser XL,
-		actions_ser XL
+	$sql = "
+		CREATE TABLE IF NOT EXISTS watcher_mail_filter (
+			id INT UNSIGNED DEFAULT 0 NOT NULL,
+			pos SMALLINT UNSIGNED DEFAULT 0 NOT NULL,
+			name VARCHAR(128) DEFAULT '' NOT NULL,
+			created INT UNSIGNED DEFAULT 0 NOT NULL,
+			worker_id INT UNSIGNED DEFAULT 0 NOT NULL,
+			criteria_ser MEDIUMTEXT,
+			actions_ser MEDIUMTEXT,
+			PRIMARY KEY (id)
+		) ENGINE=MyISAM;
 	";
-	$sql = $datadict->CreateTableSQL('watcher_mail_filter', $flds);
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute($sql);	
 }
 
 // Migrate and then drop the old forward table
@@ -136,13 +85,13 @@ if(isset($tables['worker_mail_forward'])) {
 	$sql = "SELECT id, worker_id, group_id, bucket_id, email, event FROM worker_mail_forward";
 	$rs = $db->Execute($sql);
 	
-	while(!$rs->EOF) {
-		@$old_id = intval($rs->fields['id']);
-		@$worker_id = intval($rs->fields['worker_id']);
-		@$group_id = intval($rs->fields['group_id']);
-		@$bucket_id = intval($rs->fields['bucket_id']);
-		@$email = $rs->fields['email'];
-		@$event = $rs->fields['event'];
+	while($row = mysql_fetch_assoc($rs)) {
+		@$old_id = intval($row['id']);
+		@$worker_id = intval($row['worker_id']);
+		@$group_id = intval($row['group_id']);
+		@$bucket_id = intval($row['bucket_id']);
+		@$email = $row['email'];
+		@$event = $row['event'];
 		
 		$name = 'Notification';
 		$criteria = array();
@@ -204,13 +153,12 @@ if(isset($tables['worker_mail_forward'])) {
 		
 		// Delete source row (for partial success)
 		$db->Execute(sprintf("DELETE FROM worker_mail_forward WHERE worker_id=%d AND id=%d", $worker_id, $old_id));
-		
-		$rs->MoveNext();
 	}
 	
+	mysql_free_result($rs);
+	
 	// Drop old table
-	$sql = $datadict->DropTableSQL('worker_mail_forward');
-	$datadict->ExecuteSQLArray($sql);
+	$db->Execute('DROP TABLE worker_mail_forward');
 }
 
 // ===========================================================================
